refactor: package rename data_collector to aggregated_collector
