+++ eox_core/api/data/aggregated_collector/v1/permissions.py
@@ -1,5 +1,5 @@
 """
-Custom permission classes for the Data Collector API.
+Custom permission classes for the Aggregated Collector API.
 """
 
 import logging
@@ -11,14 +11,14 @@ from rest_framework.permissions import BasePermission
 logger = logging.getLogger(__name__)
 
 
-class DatacollectorPermission(BasePermission):
+class AggregatedCollectorPermission(BasePermission):
     """
     Permission class to allow access only if the request contains a valid GitHub Action token.
     """
 
     def has_permission(self, request, view):
         auth_header = get_authorization_header(request).decode('utf-8')
-        auth_token = settings.EOX_CORE_DATA_COLLECT_AUTH_TOKEN
+        auth_token = settings.EOX_CORE_AGGREGATED_COLLECT_AUTH_TOKEN
         if auth_header and auth_header == f"Bearer {auth_token}":
             return True
         return False
